requirement.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.6
Pillow==8.3.1
pytz==2021.1
qrcode==7.2
sqlparse==0.4.1
